+++ example/lib/main.dart
@@ -59,7 +59,7 @@ class _MyHomePageState extends State<MyHomePage> with WidgetsBindingObserver {
       'name': 'Objective-C Programming (Pre-Course Workbook',
       'link':
           'https://www.bignerdranch.com/documents/objective-c-prereading-assignment.pdf'
-    }
+    },
   ];
 
   final _images = [
@@ -89,12 +89,12 @@ class _MyHomePageState extends State<MyHomePage> with WidgetsBindingObserver {
     {
       'name': 'Big Buck Bunny',
       'link':
-          'https://www.sample-videos.com/video/mp4/720/big_buck_bunny_720p_30mb.mp4'
+          'http://commondatastorage.googleapis.com/gtv-videos-bucket/sample/BigBuckBunny.mp4'
     },
     {
-      'name': 'Star Trails',
+      'name': 'Elephant Dream',
       'link':
-          'http://mirrors.standaloneinstaller.com/video-sample/star_trails.mp4'
+          'http://commondatastorage.googleapis.com/gtv-videos-bucket/sample/ElephantsDream.mp4'
     }
   ];